Fix dateStart/dateEnd
All dates are in ISO format with UTC offset info.

Session dates (session['meta'] and session['data'][*]['session'])
are based on info from proceedings (sitzung-start-uhrzeit/sitzung-ende-uhrzeit)
which is in local time. UTC offset is derived from media feed (which has the info).

For each speech, speech['dateStart']/speech['dateEnd'] are
the precise timing of the speech from the media feed.

Closes #4
Closes #5